feat(updated): add useHome & useCategoriesList hooks
