+++ src/screens/CategoriesList/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useCategoriesList} from './useCategoriesList';
