+++ src/screens/Home/Home.tsx
@@ -1,84 +1,48 @@
 import React, {useEffect, useCallback} from 'react';
-import {RefreshPageReminder, SuspenseView, useToast} from 'atoms';
+import {RefreshPageReminder, SuspenseView} from 'atoms';
 import {HomeSectionBar} from 'organisms';
-import {
-  AppStateStatus,
-  FlatList,
-  InteractionManager,
-  RefreshControl,
-  View,
-} from 'react-native';
+import {FlatList, InteractionManager, RefreshControl, View} from 'react-native';
 
-import {themeStyles} from './styles';
-import {
-  getHomeDataUpdatesRequest,
-  getInitialHomeDataRequest,
-  getHomeDataUpdateAvailableRequest,
-} from 'core/reducers';
-import {useDispatch, useSelector} from 'react-redux';
-import {selectHomeData, selectIsUpdatesAvailable} from 'core/selectors';
-import {
-  useAppState,
-  useRequestError,
-  useRequestLoading,
-  useTranslation,
-  useColorScheme,
-  useThemeStyles,
-  useHomeAnalytics,
-} from 'core/hooks';
-import {IProps} from './types';
+import {getHomeDataUpdateAvailableRequest} from 'core/reducers';
+
+import {useAppState} from 'core/hooks';
 import {COLORS} from 'assets';
-import {useFocusEffect, useIsFocused} from '@react-navigation/core';
+import {useFocusEffect} from '@react-navigation/core';
 import {ErrorToast} from '../../components/molecules';
 import {screenOptions} from './screenOptions';
-import {IObject, ITransformedCategory} from 'core/types';
+import {useHome} from './hooks';
 
-export const Home = ({navigation: {navigate}}: IProps) => {
-  const {t} = useTranslation('home');
-  const styles = useThemeStyles(themeStyles);
-  const dispatch = useDispatch();
-  const theme = useColorScheme();
-  const homeData = useSelector(selectHomeData);
-  const isUpdatesAvailable = useSelector(selectIsUpdatesAvailable);
-  const {ref, show: showToast} = useToast();
-  const isFocused = useIsFocused();
-
+export const Home = () => {
   const {
+    t,
+    updateError,
+    showToast,
+    checkUpdates,
+    dispatch,
+    loading,
+    homeData,
+    error,
     listRef,
-    sendSelectCardEvent,
-    sendSelectAllEvent,
-    sendSaveCardEvent,
-    sendUnsaveCardEvent,
-  } = useHomeAnalytics();
-
-  const getData = useCallback(() => {
-    dispatch(getHomeDataUpdatesRequest());
-  }, [dispatch]);
-
-  const getInitialData = useCallback(() => {
-    dispatch(getInitialHomeDataRequest());
-  }, [dispatch]);
-
-  const {loading} = useRequestLoading(getInitialHomeDataRequest);
-  const {error} = useRequestError(getInitialHomeDataRequest);
-
-  const {loading: refreshing} = useRequestLoading(getHomeDataUpdatesRequest);
-  const {error: updateError} = useRequestError(getHomeDataUpdatesRequest);
+    getInitialData,
+    styles,
+    theme,
+    refreshing,
+    isFocused,
+    getData,
+    navigateToObjectDetails,
+    onCategoryPress,
+    onAllObjectsPress,
+    navigateToCategoriesList,
+    sendIsFavoriteChangedEvent,
+    isUpdatesAvailable,
+    ref,
+  } = useHome();
 
   useEffect(() => {
     if (updateError) {
       showToast();
     }
   }, [showToast, updateError]);
-
-  const checkUpdates = useCallback(
-    (state: AppStateStatus, prevState: AppStateStatus) => {
-      if (state === 'active' && prevState === 'background') {
-        dispatch(getHomeDataUpdateAvailableRequest());
-      }
-    },
-    [dispatch],
-  );
 
   useAppState(checkUpdates);
 
@@ -93,55 +57,6 @@ export const Home = ({navigation: {navigate}}: IProps) => {
     ),
   );
 
-  const navigateToObjectsList = useCallback(
-    ({categoryId, title}: {categoryId: string; title: string}) => {
-      navigate('ObjectsList', {categoryId, title});
-    },
-    [navigate],
-  );
-
-  const onCategoryPress = useCallback(
-    (category: ITransformedCategory, parentCategoryName: string) => {
-      navigateToObjectsList({categoryId: category.id, title: category.name});
-      sendSelectCardEvent(category.name, parentCategoryName);
-    },
-    [navigateToObjectsList, sendSelectCardEvent],
-  );
-
-  const onAllObjectsPress = useCallback(
-    (data: {categoryId: string; title: string}) => {
-      sendSelectAllEvent(data.title);
-      navigateToObjectsList(data);
-    },
-    [navigateToObjectsList, sendSelectAllEvent],
-  );
-
-  const navigateToCategoriesList = useCallback(
-    ({categoryId, title}: {categoryId: string; title: string}) => {
-      navigate('CategoriesList', {categoryId, title});
-      sendSelectAllEvent(title);
-    },
-    [navigate, sendSelectAllEvent],
-  );
-
-  const navigateToObjectDetails = useCallback(
-    ({id, name, category}: IObject) => {
-      navigate('ObjectDetails', {objectId: id});
-      sendSelectCardEvent(name, category.name);
-    },
-    [navigate, sendSelectCardEvent],
-  );
-
-  const sendIsFavoriteChangedEvent = useCallback(
-    ({name, category}: IObject, nextIsFavoriteStatus: boolean) => {
-      if (nextIsFavoriteStatus) {
-        sendSaveCardEvent(name, category.name);
-      } else {
-        sendUnsaveCardEvent(name, category.name);
-      }
-    },
-    [sendSaveCardEvent, sendUnsaveCardEvent],
-  );
   return (
     <View style={{overflow: 'hidden', flex: 1}}>
       <SuspenseView
