+++ src/screens/Home/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useHome} from './useHome';
